chore: work-in-progress refactoring the portfolio construction model (2)
