+++ config.json
@@ -198,5 +198,6 @@
                 "sort": 89
             }
         ]
-    }
+    },
+    "python-venv": 1
 }
